Harmonized word boundaries exposed through accessibility APIs with the words produced by cursor navigation
When using a screen reader and the user is moving by word throughout a webpage,
both in editable and in non-editable content, the assistive software should announce
either the word the user's cursor moved over or the word to the right of the cursor, depending on
platform, and using accessibility APIs exposed by Chromium.
To navigate in editable content the user uses Chromium's word navigation
functionality, such as Ctrl+Left/Right arrow or Option-Left/Right arrow.
In non-editable content, the user relies on screen reader shortcuts, which in turn rely on the same
accessibility APIs.

The bug is caused by the fact that commands such as Ctrl+Left/Right do not use the same word boundaries as
the accessibility APIs.

This patch harmonizes the two sets of boundaries. To do this, I used and was guided by the following:
1. ICU's word break iterator whose rules are found at
http://userguide.icu-project.org/boundaryanalysis
and
http://www.unicode.org/reports/tr29/#Word_Boundaries
2. Idiosyncrasies of the Blink editing code which are based on established platform behavior
that Chromium needs to replicate.
//third_party/blink/renderer/core/editing/selection_adjuster.cc
StartOfWordPosition, EndOfWordPosition.
//third_party/blink/renderer/core/editing/visible_units_word.cc
StartOfWordPositionInternal, EndOfWordPositionInternal,
NextWordPositionInternal, PreviousWordPositionInternal.

AX-Relnotes: Screen reader reads the correct word when moving forward or backward
by word and there are punctuation marks or line breaks in the text.

R=aleventhal@chromium.org, dmazzoni@chromium.org, chrishtr@chromium.org

Change-Id: I2915fb4243f74f33c31508c4fc3c9e49a57209f6
Bug: 1028830
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2237930
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783871}