Modify AppService for to handle compressed standard icons.
For web apps icon, only modify the effects for uncompressed icon,
because for compressed icons, we still need apply those icon effects.

For Arc apps, when the standard icon flag is enable, call ArcAppIcon to
generate the adaptive icon for both the compressed and uncompressed
icon, so that AppService can generate the standard compressed and
uncompressed icons.

BUG=1083331

Change-Id: Icf6381ea8de96bfdc176d4df687403b72691f426
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2272659
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783873}