[Signin] Add |Continue as| button to account picker web sign-in flow
This CL adds the |Continue as| to account picker bottom sheet in the
web sign-in flow and the collapsing/expanding mechanism to the bottom
sheet. Prior to this CL, the bottom sheet shows a straightforward
recycler view of the account list as in https://crbug.com/1081253#c1.

Screenshot: https://crbug.com/1093449#c2, https://crbug.com/1093795
Bug: 1093449
Change-Id: Ie4460523ea794d48efff384ba670c23348ae5c46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2238192
Commit-Queue: Alice Wang <aliceywang@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783892}