Change "FPS meter" to "Frame Rendering Stats"
With this CL landed:
https://chromium-review.googlesource.com/c/chromium/src/+/2244002

The fps meter now have a completely new look, and it is actually
showing the throughput of the page. This CL changes the name of
"FPS meter" to "Frame Rendering Stats". Its description is also
updated.

The corresponding devtool front end changes is here
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2269939

Bug: None
Change-Id: I22cc52abd37e12bb80db99daec339d467d4595f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2268398
Commit-Queue: Xida Chen <xidachen@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783896}