COOP: rename is_initiator_aboutblank
In the function:
ShouldSwapBrowsingInstanceForCrossOriginOpenerPolicy(...)

Rename the parameter:
  is_initiator_aboutblank
into:
  is_initial_navigation

The previous name was misleading. What we want to represent is whether
this navigation is from the initial empty document or not. It doesn't
really depend on the URL of the initiator.

Note: There are some issues with callers of this function as well,
is_initial_navigation is currently set to:
!current_rfh->has_committed_any_navigation().
This gives the expected boolean most of time, except when:
1) The |current_rfh| is not the empty document and is crashed.
2) The |current_rfh| is the new RenderFrameHost, that committed before
   the navigation, due to an optimization after crash.
Those 2 issues will be handled in followups.

Bug: chromium:1076879
Change-Id: I1512aa91a3b46bdde25b36587ab9ae4f4941fa46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2273139
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Arthur Hemery <ahemery@chromium.org>
Reviewed-by: Pâris Meuleman <pmeuleman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783906}