[iOS][Password Check] Make IOSChromePasswordCheckManager alive during check
This change allows IOSChromePasswordCheckManager to exist independently
from UI when password check is running. If check is finished and no UI
holds a pointer to it, the object will be destroyed. Also a factory to
create IOSChromePasswordCheckManager was added. This factory holds only
weak pointer to the manager.

Bug: 1075494
Change-Id: Ib617b03c8ca4cec70924caa7c7e9a17deb090fbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2263814
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Viktor Semeniuk <vsemeniuk@google.com>
Cr-Commit-Position: refs/heads/master@{#783911}