[COOP] access reporting [1/N] List the accesses.
All of this is put behind a flag disabled by default.

This is mostly based on the initial prototype:
https://chromium-review.googlesource.com/c/chromium/src/+/2223934/24

This patches list all the potential accesses to be checked and reported.
The CrossOriginOpenerPolicyReporter is preparing itself to install all
the CoopAccessMonitor to the renderer(s), but It doesn't send them for
now.

To write a meaningful patch, this will need accesses to COOP-Report-Only
headers and the virtual browsing context group.

Explainer [WIP]:
https://github.com/camillelamy/explainers/blob/master/coop_reporting.md

Specification [WIP]:
https://github.com/whatwg/html/pull/5518

Tests [WIP]:
https://wpt.fyi/results/html/cross-origin-opener-policy/access-reporting

Doc [WIP]:
https://docs.google.com/document/d/1H8Be0w27fKPXKqyuJj9oEqIJEjB9Rw5AP3x-w-Fx2Zg/edit?usp=sharing

Bug: chromium:1090273
Change-Id: I5d4c613a671f99ba15b4b174431d8d3ddeaa44c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264294
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Pâris Meuleman <pmeuleman@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783972}