Check render_accessibility_ in RenderFrameHostImpl::AccessibilityHitTest
This was previously a helper function so it didn't need a null-check,
but now it's public, so we need to check that render_accessibility_
is still valid. During teardown it's possible that
render_accessibility_ is null but the corresponding
BrowserAccessibilityManager still exists and could try to
call AccessibilityHitTest.

Bug: 1096265
Change-Id: I353b449ec44ec3adc7cd60d4615148505c13efa4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2271106
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783982}