LocalDOMWindow::OutgoingReferrer() should tolerate GetFrame() returning null
When I moved the implementation of OutgoingReferrer() from Document to
LocalDOMWindow in
https://chromium-review.googlesource.com/c/chromium/src/+/2258426, I
removed a GetFrame() null-check.

https://chromium-review.googlesource.com/c/chromium/src/+/2258955
uncovered that this because in certain situations we will give
FrameLoadRequest a LocalDOMWindow that has been detached. This appears
to happen when a window schedules a FormSubmission (presumably targeted
at a different window), then is detached before the FormSubmission
executes asynchronously. I haven't been able to actually trigger this
in a test, though.

Bug: 1100002
Change-Id: I5231d8cd33e03736c0fd6bad9bd490347b9662b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2273758
Auto-Submit: Nate Chapin <japhet@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783986}