Enable feature kUseNSURLSessionForGaiaSigninRequests by default
The dependency bug on Gaia infrastructure (b/157015322) was fixed so it
is now safe to enable feature kUseNSURLSessionForGaiaSigninRequests by
default.

This CL enables this feature by default. The finch kill switch is kept
to ensure that we can disable the feature in case of any regression.

Bug: 944442
Change-Id: I8a4bcfecd0d7816fd4b00308ab9c925341dbeb32
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247023
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784019}