Drop shell browser web test code from the Android build.
In order to add a WebTestShellPlatformDelegate override for other
platforms, I need to add one for each web test build configuration to
avoid messy defines in the header. That would also need an Android
version even though we don't run web tests there.

So in this CL we group all the web test files/deps together in the
BUILD.gn and do not include them on android. This allows us to avoid
adding a WebTestShellPlatformDelegate implementation for the platform.

Moves the MockClientHintsControllerDelegate class up from web_test/
to the content/test/ directory, as it is shared test code between
content_browsertests and web tests. We also move its dependent
mock_client_hints_utils.{h,cc} helpers with it to content/test/.

This removes the 2 dependencies from content_browsertests on web test
code via content shell, now that the web test code will not be present
on Android, but content_browsertests are.

R=avi@chromium.org

Change-Id: I7f40fc67f4a6c993dd7eccac0da5bc7482806fe3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2270800
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784089}