[cc] Add  TASK_CATEGORY_BACKGROUND_WITH_NORMAL_THREAD_PRIORITY.
The new task category is for tasks that must not be scheduled when
there are foreground tasks to run, but shouldn't run at background
thread priority. Tasks that acquire resources shared with foreground
work (example: a lock) should not run at background thread priority,
because there is a risk that they will hold these resources for
a long time, causing a priority inversion.

The new task category is used in https://crrev.com/c/2240141 to fix
a priority inversion that is observed via Slow Reports.

Bug: 1072756
Change-Id: Ie42feb48cc167303712e5adbdfeec11ec2e312c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2158939
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Commit-Queue: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784092}