Keyboard up/down changes destinations when dropdown is closed
This CL creates a dedicated function for handling up/down key presses
when focus is on the print preview dropdown.

To match the <select> element's behavior, when at the bottom of the
dropdown list, pressing down will not wrap back to the top. And vice
versa.

Bug: 1059607
Change-Id: I89d977617745ba2412eeed08f0312fc6f0f01d4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2273497
Commit-Queue: Gavin Williams <gavinwill@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784120}