Play resampled audio until EOS
When nearing the EOS using a resampler to play audio, we sometimes omit
the last few frames. This happens when we send all remaining audio
buffers to the resampler and partially fill it with silence, but we
don't know if we have completely consumed the non-silence frames.

This CL fixes the issue by first checking that we have completely filed
the resampler with silence, before declaring there is nothing else to
play. By doing so, we play silence after the last valid frames, but we
will at least play all valid frames.

Change-Id: I8ea4d504907746b2ea170ca55169718aee6d8586
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276526
Auto-Submit: Thomas Guilbert <tguilbert@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784197}