[Offline indicator v2] Add Java animations for grid tab switcher
This CL adds browser controls animations for the tab switcher. As
changing the margins in every animation frame would cause a relayout and
be expensive, we use the translationY property to animate the page. The
margins are applied only once at the start or end of the animation.

Bug: 1075640
Change-Id: Ic8af8fa9e043bf24bdedcf87ce2ca793f73baa6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2269197
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784198}