Use FakeServiceWorkerContext in IsolatedPrerenderTabHelper unittests
This CL removes the ForTest() functions in service_worker_context.h,
specifically:
- WaitForRegistrationsInitializedForTest()
- AddRegistrationToRegisteredOriginsForTest(const url::Origin& origin)

The motivation of this change is to avoid unnecessarily exposing
methods from service_worker_context_wrapper.cc, and use the
FakeServiceWorkerContext instead of using the real context when it is
not needed. This leads to better consistency with the storage as well.

Bug: 807440
Change-Id: I01791b82062d8e3c3bef63b029e6f77d1fe02469
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2262398
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Reviewed-by: Patrick Monette <pmonette@chromium.org>
Commit-Queue: Nidhi Jaju <nidhijaju@google.com>
Cr-Commit-Position: refs/heads/master@{#784201}