[NativeFS] Create a new NativeFileSystemTransferToken implementation
In order to implement Drag & Drop, extend the abstract base class
NativeFileSystemTransferToken with an implementation that's not tightly
coupled to an origin. When exchanging this token for an
NativeFileSystemFileHandle/NativeFileSystemDirectoryHandle, a caller
should pass in a BindingContext and the NativeFileSystemManager will
lazily create the handle.

Bug: 1080811
Change-Id: If6daaf628ec27d2e23339901d3db22d9c0386e08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2243997
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Darwin Huang <huangdarwin@chromium.org>
Commit-Queue: Hazim Mohamed <hazimmohamed@google.com>
Cr-Commit-Position: refs/heads/master@{#784217}