media: Add MediaFoundationCdmSession
This class adapts Media Foundation IMFContentDecryptionModuleSession
to a class that uses Chromium CDM types. It also handles setting session
ID which isn't available when the session is created. This makes it
easier for MediaFoundationCdm (to be uploaded in the next CL) to manage
multiple sessions.

NOPRESUBMIT=true

Bug: 999747
Change-Id: Id966915d8bebb97406a412c4d2e958f2e2529576
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2265099
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: John Rummell <jrummell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784251}