Download later: Change the title of the location dialog.
The location dialog title is changed to "Choose where to download" to
match the download later dialog's "Choose when to download".

The file size is kept, the text could be "Choose where to download 15Kb"
when the file size is available.

Also does layout polish to make the title and subtitle to match the
new download later dialog.

Bug: 1101010,1078454
Change-Id: I08c4965ca007afa74b5c23e019d615fbb8fcbc81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276347
Commit-Queue: Xing Liu <xingliu@chromium.org>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Hesen Zhang <hesen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784266}