[FragmentItem] Include floats in associated LayoutObject
This patch changes to associate floating |LayoutObject|s to
|NGFragmentItem| using |FirstInlineFragmentItemIndex|.

The index is used for many purposes, but one of them is to
mark destroyed or moved |LayoutObject| in |NGFragmentItem|.
Because floating objects are stored in |NGFragmentItem|,
they need to be indexed to mark correctly.

Also changes |LayoutBlockFlow::
SetShouldDoFullPaintInvalidationForFirstLine| to skip items
for destroyed or moved |LayoutObject| because it may be
called on dirty tree.

Bug: 1100900
Change-Id: If35bf2c77b9a46a9ba598e6785926f6b0e1c5455
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275373
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Aleks Totic <atotic@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784297}