[User menu] Refactor setting profile identity into one function
This CL is a refactoring before a minor update of the design of the
profile menu. This CL brings no behavioral change.

This change is needed because in the new design, the layout of the
profile name heading and the layout of the identity info are more
intertwined and thus it is opportune to set them in one function.

Bug: 1099286
Change-Id: Iad6f03bd3ff7fa1e17cc1cb61b2d76c952c1394f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2274626
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Auto-Submit: Jan Krcal <jkrcal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784325}