[InfoBars] Componentize InfoBarContainerLayout
This class can be reused as-is by WebLayer. To facilitate the
componentization, we break the dependency on InfoBarContainer by moving
the InfoBarAnimationListener interface into its own file in
//components/infobars as well.

InfoBarWrapper will be able to return to being package-private once I
change //weblayer to use the componentized InfoBarContainerLayout in a
followup. Whether InfoBarContainerLayout's methods will be able to go
back to being package-private depends on the resolution of sharing
InfoBarContainerView, which I will be examining shortly.

Bug: 1099169
Change-Id: I50b68b8269dd7f0df12fe4f8bda1c32baf43e55b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275873
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784353}