[WebLayer] Reuse componentized InfoBarContainerLayout
//chrome's InfoBarContainerLayout.java has now been componentized and
WebLayer can reuse it. As part of that componentization, the
InfoBarAnimationListener interface was pulled out of //chrome's
InfoBarContainer.java to live in a standalone //components-level file.
This CL changes WebLayer to use the componentized
InfoBarContainerLayout and the componentized InfoBarAnimationListener
on which the former depends. There are no behavioral changes, as none
of the code being deleted was changed from the now-componentized
//chrome versions.

Bug: 1099169
Change-Id: Ic443dc5cd266da1add574e85caf100a809111e2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275895
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784365}