Safety check on Android: implemented update check and integrated with Omaha
Currently Omaha code for Android is in //chrome/android, while Safety check is in //chrome/browser.
In order for Safety check to be able to use Omaha, there is a need to have a delegate Omaha client class with an interface in //chrome/browser and an implementation in //chrome/android.
This CL also adds all the associated update logic to Safety check on Clank.

Bug: 1070620
Change-Id: I3119cafe23ccb9cbc5b49edbcc7895576e77cefd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264442
Commit-Queue: Andrey Zaytsev <andzaytsev@google.com>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Auto-Submit: Andrey Zaytsev <andzaytsev@google.com>
Cr-Commit-Position: refs/heads/master@{#784370}