Cert Provisioning: Ignore certificates without attributes
Ignore certificates that return error when asked for
certificate provisioning specific attribute. The error
most likely just means that the certificate doesn't
have a value for the attribute. This will be changed
later when there is a way to distinguish unset
attribute case and other errors.

Bug: 1045895
Test: CertProvisioning*
Change-Id: Ib4960916f53e50ed6d374085cbd165b8fdcd4c68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275879
Commit-Queue: Michael Ershov <miersh@google.com>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784422}