Fix app list sync paging issue caused by incompatible app overflow.
When there are incompatible apps on different type of devices under the
same user account, it is possible that moving or adding an app on an
empty spot on a page of a different type of device (e.g. Device 1) may
cause app overflow on another device (e.g. Device 2) since it may have
more apps on the same page. See details in
http://crbug.com/1098174#c11.

When the change is synced to the Device 2, paged view structure may load
meta data and detect a full page of apps without a page break item at
the end of the overflowed page, which will mess the calculation in
PagedViewStructure::GetTargetItemIndexForMove later if user moves app
around. Therefore, after the sync service has finished processing sync
change, SaveToMetaData should be called to insert a page break item at
the end of each full page if the page break is found missing.

Bug: 1098174
Change-Id: Iccbfde81810e337662fea2bec71e25d53d8b08db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2277201
Commit-Queue: Jenny Zhang <jennyz@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784546}