Trust Tokens: Clean up browser tests before adding multiple-origin tests
In order to make it easier to extend the Trust Tokens browser tests to
cover more cases, including some sequences of issuance and redemption
operations with multiple origins involved (cross-site issuance, for
instance, or a sequence of issuance operations against different issuing
origins), this CL refactors the test server to use CERT_TEST_NAMES so
that it's possible to connect via https to a number of test domains,
like a.test and b.test. Unfortunately, this means that it's no longer
possible to connect via https to good old 127.0.0.1, so the existing
tests change to use a.test instead.

Related changes necessary for this:
- stop looking at HttpRequest::GetURL when verifying signed requests;
rather, infer the request's destination domain from its Host header
- when setting the "issuer" parameter for outgoing signed requests, use
a value dependent on the host (e.g. "a.test") and the embedded test
server's port (fixed for the duration of the test, but liable to change
between tests)

This CL also includes a couple other cleanup changes, like changing all
EXPECT_TRUE(NavigateToURL(...)) to ASSERTS, so that it's as easy as
possible to tell that a test has failed to set up.

Bug: 1071293
Change-Id: Ic07646d3410b5313f61c02422f68121086c8aebb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2274174
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784576}