[LayoutNG] Fragmentation of floats in inline formatting contexts.
When breaking before or inside a float that's inside a block that
establishes an inline formatting context, we create and propagate break
tokens all the way up to the nearest block container algorithm. When
resuming the floats in the next fragmentainer, they will be treated as
normal floated children of that block, outside of the inline formatting
context. A fragment item will (exactly as normally, without block
fragmentation) be created for the fragment in the first fragmentainer in
which the float occurs. For all remaining fragmentainers, however, only
a regular physical box fragment child will be created for the float. No
inline layout will be involved for the sake of resumed floats.

This approach isn't entirely correct, e.g. if there's a non-atomic
inline parent that causes a paint effect (filter, relative offset,
etc.), only the first fragment will be affected by such parents. What
remains pretty clear, though, is that we cannot re-enter regular inline
formatting context layout for resumed floats. The parents may have ended
in a previous fragmentainer or line, so we cannot just continue inside
them in the next fragmentainer. We probably somehow want to wrap the
resumed floats inside fragments created for the non-atomic inlines that
affect them.

This implementation detail requires some changes in painting and
hit-testing of resumed floats, so that we're able to walk *both*
fragment items and physical box fragment children. IsLastForNode()
determination for floating items also needs an update, since the float
will only be a fragment item in the fragmentainer that it's initially
defined, and just a regular physical box fragment child in all
subsequent fragmentainers. The fix is simple: Just check the outgoing
break token instead.

Re-enable and update some unit tests that got disabled back when IFC
float fragmentation support got removed. It should be mentioned, though,
that these tests aren't that useful anymore, since the floats for the
most part are "hidden" as fragment items inside the line box fragments.
Luckily, we're now able to run web tests with NG block fragmentation,
and this CL fixes 22 existing tests.

Bug: 915929, 829028
Change-Id: I4e869341a05ea8778e4ce8204b7fbc103863868a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264367
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#784607}