GestureNav: Let the Coordinator own NavigationHandler
Main classes of gesture navigation feature are:

- HistoryNavigationCoordinator: self-explanatory

- HistoryNavigationLayout: Top-level view. Added to
  CompositorViewHolder, and encloses SideSlideLayout which has arrow
  puck and overscroll glow as children. SideSlideLayout gets attached to
  view hieararchy only when in use.

- NavigationHandler: Contains main logic that handles user
  gesture/UI control.

The goal is to make NavigationHandler play the role of
Mediator. Currently it is created and managed by
HistoryNavigationLayout, which undesirably introduced lots of logic
inside the view. This CL moves most of the logic out of the view, and
let the coordinator create and manage the handler instead.

This change helps slim down the view to pure Android view (the UI is
simple, therefore doesn't require view model), and make overall
structure close to MVC.

Bug: 1100717
Change-Id: I2985f341d93785248db80221d297081cbfd9d1b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275279
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784626}