Add a first pass of the desktop test wrapper scripts.
As described in the linked bug and documented in
https://docs.google.com/document/d/19qpDeR2z-8j-WQJC-5waSd134kz94UXgYU_e_N-VplU/edit?usp=sharing
we want to change how tests are launched under swarming so that
they all have a consistent API and usage.

We need to do at least some of this work now in order to unblock
work going on in Swarming to change how a task's command line and
list of files are specified, but doing all of the work in the linked
doc and bug will significantly restructure how we configure and launch
tests, and should bring a lot more consistency to our continuous
integration infrastructure.

This CL is a first step: a proof of concept that converts
url_unittests on desktop to the new API in a backwards-compatible
way (i.e., this will work with the existing chromium recipes).

R=tikuta@chromium.org, martiniss@chromium.org, bpastene@chromium.org, estaab@chromium.org, kbr@chromium.org

Bug: 816629
Change-Id: I4aaaed9853e451536f21a752d896cf7bef5fff18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2249173
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Erik Staab <estaab@chromium.org>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784653}