[Suggested files] Fixes for multiprofile
Currently, file results are sometimes not displayed on the first
launcher open after switching to a user for the first time via
multiprofile. This is due to when we decide to warm the Drive results
cache: when DriveFS is first mounted. For a single user, this happens
soon after the user logs in. However when switching between multiprofile
users, drivefs is not re-mounted on each switch so the cache isn't
warmed.

This is fixed by checking whether drivefs is already mounted on
construction and, if so, just fetching results immediately.

Bundled with the change, I've also reduces the minimum time between
saves for the zero state files model, because the 300s window drops too
many actions  in a multiprofile setting.

Bug: 1101107
Change-Id: I418e1cecca3d92279f237bc618e9fe1557c7b331
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275930
Reviewed-by: Thanh Nguyen <thanhdng@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784660}