[Autofill Assistant] First step refactor |ClickOrTap|
Refactor |ClickOrTap| to take |ElementFinder::Result| reference instead
of a |Selector|.

The |FindElement| logic before calling the |ClickOrTap| has been
extracted into the |action_delegate_util|. This is supposed to be
used by the |ActionDelegate| callers (ClickAction and
RequiredFieldsFallbackHandler).

Since |ElementFinder| returns a unique_ptr and the |ClickOrTap| takes a
raw pointer, the object needs to be kept alive until the end by passing
it along to the final callback.

Bug: b/158153191
Change-Id: I832dd32a79c12d3246b9118613a0321eae50020d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2230458
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#784711}